@@ -3,47 +3,46 @@ using CodaParser.LineParsers;
 using CodaParser.Lines;
 using NUnit.Framework;
 
-namespace CodaParserTests.LineParsers
+namespace CodaParserTests.LineParsers;
+
+[TestFixture]
+public class InitialStateLineParserTest
 {
-    [TestFixture]
-    internal class InitialStateLineParserTest
+    [Test]
+    public void TestAccountIsIbanIsSetCorrectly()
     {
-        [Test]
-        public void TestAccountIsIbanIsSetCorrectly()
-        {
-            var parser = new InitialStateLineParser();
+        var parser = new InitialStateLineParser();
 
-            var sample = "13155001548226815 EUR0BE                  0000000004004100241214CODELICIOUS               PROFESSIONAL ACCOUNT               255";
+        var sample = "13155001548226815 EUR0BE                  0000000004004100241214CODELICIOUS               PROFESSIONAL ACCOUNT               255";
 
-            Assert.IsTrue(parser.CanAcceptString(sample));
+        Assert.That(parser.CanAcceptString(sample), Is.True);
 
-            var result = (InitialStateLine)parser.Parse(sample);
+        var result = (InitialStateLine)parser.Parse(sample);
 
-            Assert.AreEqual(true, result.Account.Number.IsIbanNumber);
-        }
+        Assert.That(result.Account.Number.IsIbanNumber, Is.EqualTo(true));
+    }
 
-        [Test]
-        public void TestSample1()
-        {
-            var parser = new InitialStateLineParser();
+    [Test]
+    public void TestSample1()
+    {
+        var parser = new InitialStateLineParser();
 
-            var sample = "10155001548226815 EUR0BE                  0000000004004100241214CODELICIOUS               PROFESSIONAL ACCOUNT               255";
+        var sample = "10155001548226815 EUR0BE                  0000000004004100241214CODELICIOUS               PROFESSIONAL ACCOUNT               255";
 
-            Assert.IsTrue(parser.CanAcceptString(sample));
+        Assert.That(parser.CanAcceptString(sample), Is.True);
 
-            var result = (InitialStateLine)parser.Parse(sample);
+        var result = (InitialStateLine)parser.Parse(sample);
 
-            Assert.AreEqual("0", result.Account.NumberType.Value);
-            Assert.AreEqual(155, result.StatementSequenceNumber.Value);
-            Assert.AreEqual("001548226815", result.Account.Number.Value);
-            Assert.AreEqual("EUR", result.Account.Currency.CurrencyCode);
-            Assert.AreEqual("BE", result.Account.Country.CountryCode);
-            Assert.AreEqual(false, result.Account.Number.IsIbanNumber);
-            Assert.AreEqual(4004.100, result.Balance.Value);
-            Assert.AreEqual(new DateTime(2014, 12, 24), result.Date.Value);
-            Assert.AreEqual("CODELICIOUS", result.Account.Name.Value);
-            Assert.AreEqual("PROFESSIONAL ACCOUNT", result.Account.Description.Value);
-            Assert.AreEqual(255, result.PaperStatementSequenceNumber.Value);
-        }
+        Assert.That(result.Account.NumberType.Value, Is.EqualTo("0"));
+        Assert.That(result.StatementSequenceNumber.Value, Is.EqualTo(155));
+        Assert.That(result.Account.Number.Value, Is.EqualTo("001548226815"));
+        Assert.That(result.Account.Currency.CurrencyCode, Is.EqualTo("EUR"));
+        Assert.That(result.Account.Country.CountryCode, Is.EqualTo("BE"));
+        Assert.That(result.Account.Number.IsIbanNumber, Is.EqualTo(false));
+        Assert.That(result.Balance.Value, Is.EqualTo(4004.100));
+        Assert.That(result.Date.Value, Is.EqualTo(new DateTime(2014, 12, 24)));
+        Assert.That(result.Account.Name.Value, Is.EqualTo("CODELICIOUS"));
+        Assert.That(result.Account.Description.Value, Is.EqualTo("PROFESSIONAL ACCOUNT"));
+        Assert.That(result.PaperStatementSequenceNumber.Value, Is.EqualTo(255));
     }
 }
