@@ -3,147 +3,146 @@ using System.Linq;
 using CodaParser;
 using NUnit.Framework;
 
-namespace CodaParserTests
+namespace CodaParserTests;
+
+[TestFixture]
+public class ParserTest
 {
-    [TestFixture]
-    public class ParserTest
+    [Test]
+    public void TestHas4EntriesWithStructuredMessage()
     {
-        [Test]
-        public void TestHas4EntriesWithStructuredMessage()
+        var parser = new Parser();
+        var result = parser.ParseFile(GetSamplePath("sample1.cod")).ToList();
+
+        Assert.That(result, Has.Count.EqualTo(1));
+        var firstResult = result.First();
+        Assert.That(firstResult.InitialBalance, Is.EqualTo(17752.12));
+        Assert.That(firstResult.NewBalance, Is.EqualTo(17832.12));
+        Assert.That(firstResult.Date, Is.EqualTo(new DateTime(2017, 10, 11)));
+        Assert.That(firstResult.InformationalMessage, Is.Empty);
+
+        Assert.That(firstResult.Transactions, Has.Count.EqualTo(4));
+        var firstTransaction = firstResult.Transactions.First();
+        Assert.That(firstTransaction.Message, Is.EqualTo("GROTE WEG            32            3215    HASSELT"));
+        Assert.That(firstTransaction.StructuredMessage, Is.EqualTo("000003505158"));
+        Assert.That(firstTransaction.Amount, Is.EqualTo(5));
+        Assert.That(firstTransaction.Account.Name, Is.EqualTo("KLANT1 MET NAAM1"));
+        Assert.That(firstTransaction.Account.Number, Is.EqualTo("BE22313215646432"));
+    }
+
+    [Test]
+    public void TestMessageOnMultipleLinesInformationBlock()
+    {
+        var parser = new Parser();
+        var result = parser.ParseFile(GetSamplePath("sample4.cod"));
+
+        Assert.That(result.First().Transactions.First().Message, Is.EqualTo("Europese overschrijving (zie bijlage)  + 17.233,54Van: COMPANY BLABLABLAH BVBA - BE64NOT PR"));
+    }
+
+    [Test]
+    public void TestMessageOnMultipleLinesTransactionBlock()
+    {
+        var parser = new Parser();
+        var result = parser.ParseFile(GetSamplePath("sample3.cod"));
+
+        Assert.That(result.First().Transactions.First().Message, Is.EqualTo("Message goes here and continues here or here"));
+    }
+
+    [Test]
+    public void TestNoAccount()
+    {
+        var parser = new Parser();
+        var result = parser.ParseFile(GetSamplePath("sample2.cod"));
+
+        Assert.That(result.First().Transactions.First().Account.Name, Is.Empty);
+        Assert.That(result.First().Transactions.First().Message, Is.EqualTo("Zichtrekening nr  21354598  - 2,11Justification in annex 00001680"));
+    }
+
+    [Test]
+    public void TestSample5SimpleFormat()
+    {
+        var parser = new Parser();
+        var result = parser.ParseFile(GetSamplePath("sample5.cod")).ToList();
+
+        Assert.That(result, Has.Count.EqualTo(1));
+        var statement = result.First();
+        Assert.That(statement.Date, Is.Not.EqualTo(default(DateTime)));
+        Assert.That(statement.Account, Is.Not.Null);
+        Assert.That(statement.InitialBalance, Is.Not.EqualTo(default(decimal)));
+        Assert.That(statement.NewBalance, Is.Not.EqualTo(default(decimal)));
+
+        Assert.That(statement.Transactions, Has.Count.EqualTo(3));
+
+        foreach (var transaction in statement.Transactions)
         {
-            var parser = new Parser();
-            var result = parser.ParseFile(GetSamplePath("sample1.cod"));
-
-            Assert.That(result, Has.Count.EqualTo(1));
-            var firstResult = result.First();
-            Assert.AreEqual(17752.12, firstResult.InitialBalance);
-            Assert.AreEqual(17832.12, firstResult.NewBalance);
-            Assert.AreEqual(new DateTime(2017, 10, 11), firstResult.Date);
-            Assert.IsEmpty(firstResult.InformationalMessage);
-
-            Assert.That(firstResult.Transactions, Has.Count.EqualTo(4));
-            var firstTransaction = firstResult.Transactions.First();
-            Assert.AreEqual("GROTE WEG            32            3215    HASSELT", firstTransaction.Message);
-            Assert.AreEqual("000003505158", firstTransaction.StructuredMessage);
-            Assert.AreEqual(5, firstTransaction.Amount);
-            Assert.AreEqual("KLANT1 MET NAAM1", firstTransaction.Account.Name);
-            Assert.AreEqual("BE22313215646432", firstTransaction.Account.Number);
+            Assert.That(transaction.Account, Is.Not.Null);
+            Assert.That(transaction.TransactionDate, Is.Not.EqualTo(default(DateTime)));
+            Assert.That(transaction.ValutaDate, Is.Not.EqualTo(default(DateTime)));
+            Assert.That(transaction.Message, Is.Not.Empty);
         }
 
-        [Test]
-        public void TestMessageOnMultipleLinesInformationBlock()
-        {
-            var parser = new Parser();
-            var result = parser.ParseFile(GetSamplePath("sample4.cod"));
+        Assert.That(statement.Transactions[0].TransactionSequence, Is.EqualTo(1));
+        Assert.That(statement.Transactions[0].StatementSequence, Is.EqualTo(214));
 
-            Assert.AreEqual("Europese overschrijving (zie bijlage)  + 17.233,54Van: COMPANY BLABLABLAH BVBA - BE64NOT PR", result.First().Transactions.First().Message);
-        }
+        Assert.That(statement.Transactions[1].TransactionSequence, Is.EqualTo(2));
+        Assert.That(statement.Transactions[1].StatementSequence, Is.EqualTo(214));
 
-        [Test]
-        public void TestMessageOnMultipleLinesTransactionBlock()
-        {
-            var parser = new Parser();
-            var result = parser.ParseFile(GetSamplePath("sample3.cod"));
+        Assert.That(statement.Transactions[2].TransactionSequence, Is.EqualTo(9));
+        Assert.That(statement.Transactions[2].StatementSequence, Is.EqualTo(214));
+    }
 
-            Assert.AreEqual("Message goes here and continues here or here", result.First().Transactions.First().Message);
-        }
+    [Test]
+    public void TestSample6()
+    {
+        var parser = new Parser();
+        var result = parser.ParseFile(GetSamplePath("sample6.cod")).ToList();
 
-        [Test]
-        public void TestNoAccount()
-        {
-            var parser = new Parser();
-            var result = parser.ParseFile(GetSamplePath("sample2.cod"));
+        Assert.That(result, Has.Count.EqualTo(1));
+        var statement = result.First();
+        Assert.That(statement.Account, Is.Not.Null);
+        Assert.That(statement.Transactions, Has.Count.EqualTo(3));
+        Assert.That(statement.Date, Is.EqualTo(new DateTime(2015, 1, 18)));
+        Assert.That(statement.InitialBalance, Is.EqualTo(4004.1));
+        Assert.That(statement.NewBalance, Is.EqualTo(-500012.1));
+        Assert.That(statement.InformationalMessage, Is.EqualTo("THIS IS A PUBLIC MESSAGE"));
 
-            Assert.IsEmpty(result.First().Transactions.First().Account.Name);
-            Assert.AreEqual("Zichtrekening nr  21354598  - 2,11Justification in annex 00001680", result.First().Transactions.First().Message);
-        }
+        Assert.That(statement.Account.Name, Is.EqualTo("CODELICIOUS"));
+        Assert.That(statement.Account.Bic, Is.EqualTo("GEBABEBB"));
+        Assert.That(statement.Account.CompanyIdentificationNumber, Is.EqualTo("09029308273"));
+        Assert.That(statement.Account.Number, Is.EqualTo("001548226815"));
+        Assert.That(statement.Account.CurrencyCode, Is.EqualTo("EUR"));
+        Assert.That(statement.Account.CountryCode, Is.EqualTo("BE"));
 
-        [Test]
-        public void TestSample5SimpleFormat()
-        {
-            var parser = new Parser();
-            var result = parser.ParseFile(GetSamplePath("sample5.cod"));
+        var transaction1 = statement.Transactions[0];
+        var transaction2 = statement.Transactions[1];
+        var transaction3 = statement.Transactions[2];
+        Assert.That(transaction1.Account, Is.Not.Null);
+        Assert.That(transaction1.TransactionDate, Is.EqualTo(new DateTime(2014, 12, 25)));
+        Assert.That(transaction1.ValutaDate, Is.EqualTo(new DateTime(2014, 12, 25)));
+        Assert.That(transaction1.Amount, Is.EqualTo(-767.823));
+        Assert.That(transaction1.Message, Is.EqualTo("112/4554/46812   813  ANOTHER MESSAGE  MESSAGE"));
+        Assert.That(transaction1.StructuredMessage, Is.Empty);
 
-            Assert.That(result, Has.Count.EqualTo(1));
-            var statement = result.First();
-            Assert.AreNotEqual(default(DateTime), statement.Date);
-            Assert.IsNotNull(statement.Account);
-            Assert.AreNotEqual(default(decimal), statement.InitialBalance);
-            Assert.AreNotEqual(default(decimal), statement.NewBalance);
+        Assert.That(transaction1.Account.Name, Is.EqualTo("BVBA.BAKKER PIET"));
+        Assert.That(transaction1.Account.Bic, Is.EqualTo("GEBCEEBB"));
+        Assert.That(transaction1.Account.Number, Is.EqualTo("BE54805480215856"));
+        Assert.That(transaction1.Account.CurrencyCode, Is.EqualTo("EUR"));
+        Assert.That(transaction1.TransactionSequence, Is.EqualTo(1));
+        Assert.That(transaction1.StatementSequence, Is.EqualTo(214));
 
-            Assert.That(statement.Transactions, Has.Count.EqualTo(3));
+        Assert.That(transaction2.Message, Is.EqualTo("54875"));
+        Assert.That(transaction2.StructuredMessage, Is.EqualTo("112455446812"));
+        Assert.That(transaction2.TransactionSequence, Is.EqualTo(2));
+        Assert.That(transaction2.StatementSequence, Is.EqualTo(214));
 
-            foreach (var transaction in statement.Transactions)
-            {
-                Assert.IsNotNull(transaction.Account);
-                Assert.AreNotEqual(default(DateTime), transaction.TransactionDate);
-                Assert.AreNotEqual(default(DateTime), transaction.ValutaDate);
-                Assert.IsNotEmpty(transaction.Message);
-            }
+        Assert.That(transaction3.Account.Name, Is.Empty);
+        Assert.That(transaction3.Account.Bic, Is.EqualTo("GEBCEEBB"));
+        Assert.That(transaction3.TransactionSequence, Is.EqualTo(9));
+        Assert.That(transaction3.StatementSequence, Is.EqualTo(214));
+    }
 
-            Assert.AreEqual(1, statement.Transactions[0].TransactionSequence);
-            Assert.AreEqual(214, statement.Transactions[0].StatementSequence);
-
-            Assert.AreEqual(2, statement.Transactions[1].TransactionSequence);
-            Assert.AreEqual(214, statement.Transactions[1].StatementSequence);
-
-            Assert.AreEqual(9, statement.Transactions[2].TransactionSequence);
-            Assert.AreEqual(214, statement.Transactions[2].StatementSequence);
-        }
-
-        [Test]
-        public void TestSample6()
-        {
-            var parser = new Parser();
-            var result = parser.ParseFile(GetSamplePath("sample6.cod"));
-
-            Assert.That(result, Has.Count.EqualTo(1));
-            var statement = result.First();
-            Assert.IsNotNull(statement.Account);
-            Assert.That(statement.Transactions, Has.Count.EqualTo(3));
-            Assert.AreEqual(new DateTime(2015, 1, 18), statement.Date);
-            Assert.AreEqual(4004.1, statement.InitialBalance);
-            Assert.AreEqual(-500012.1, statement.NewBalance);
-            Assert.AreEqual("THIS IS A PUBLIC MESSAGE", statement.InformationalMessage);
-
-            Assert.AreEqual("CODELICIOUS", statement.Account.Name);
-            Assert.AreEqual("GEBABEBB", statement.Account.Bic);
-            Assert.AreEqual("09029308273", statement.Account.CompanyIdentificationNumber);
-            Assert.AreEqual("001548226815", statement.Account.Number);
-            Assert.AreEqual("EUR", statement.Account.CurrencyCode);
-            Assert.AreEqual("BE", statement.Account.CountryCode);
-
-            var transaction1 = statement.Transactions[0];
-            var transaction2 = statement.Transactions[1];
-            var transaction3 = statement.Transactions[2];
-            Assert.IsNotNull(transaction1.Account);
-            Assert.AreEqual(new DateTime(2014, 12, 25), transaction1.TransactionDate);
-            Assert.AreEqual(new DateTime(2014, 12, 25), transaction1.ValutaDate);
-            Assert.AreEqual(-767.823, transaction1.Amount);
-            Assert.AreEqual("112/4554/46812   813  ANOTHER MESSAGE  MESSAGE", transaction1.Message);
-            Assert.IsEmpty(transaction1.StructuredMessage);
-
-            Assert.AreEqual("BVBA.BAKKER PIET", transaction1.Account.Name);
-            Assert.AreEqual("GEBCEEBB", transaction1.Account.Bic);
-            Assert.AreEqual("BE54805480215856", transaction1.Account.Number);
-            Assert.AreEqual("EUR", transaction1.Account.CurrencyCode);
-            Assert.AreEqual(1, transaction1.TransactionSequence);
-            Assert.AreEqual(214, transaction1.StatementSequence);
-
-            Assert.AreEqual("54875", transaction2.Message);
-            Assert.AreEqual("112455446812", transaction2.StructuredMessage);
-            Assert.AreEqual(2, transaction2.TransactionSequence);
-            Assert.AreEqual(214, transaction2.StatementSequence);
-
-            Assert.IsEmpty(transaction3.Account.Name);
-            Assert.AreEqual("GEBCEEBB", transaction3.Account.Bic);
-            Assert.AreEqual(9, transaction3.TransactionSequence);
-            Assert.AreEqual(214, transaction3.StatementSequence);
-        }
-
-        private string GetSamplePath(string sampleFile)
-        {
-            return System.IO.Path.Combine(TestContext.CurrentContext.TestDirectory, "Samples", sampleFile);
-        }
+    private string GetSamplePath(string sampleFile)
+    {
+        return System.IO.Path.Combine(TestContext.CurrentContext.TestDirectory, "Samples", sampleFile);
     }
 }
