@@ -2,6 +2,8 @@ using CodaParser.LineParsers;
 using CodaParser.Lines;
 using NUnit.Framework;
 
+namespace CodaParserTests.LineParsers;
+
 [TestFixture]
 public class InformationPart1LineParserTest
 {
@@ -11,22 +13,22 @@ public class InformationPart1LineParserTest
         var parser = new InformationPart1LineParser();
         var sample = "31000100010007500005482        004800001001BVBA.BAKKER PIET                                                                  1 0";
 
-        Assert.IsTrue(parser.CanAcceptString(sample));
+        Assert.That(parser.CanAcceptString(sample), Is.True);
 
         var result = (InformationPart1Line)parser.Parse(sample);
 
-        Assert.AreEqual(1, result.SequenceNumber.Value);
-        Assert.AreEqual(1, result.SequenceNumberDetail.Value);
-        Assert.AreEqual("0007500005482", result.BankReference.Value);
-        Assert.AreEqual("0", result.TransactionCode.Type.Value);
-        Assert.AreEqual("04", result.TransactionCode.Family.Value);
-        Assert.AreEqual("80", result.TransactionCode.Operation.Value);
-        Assert.AreEqual("000", result.TransactionCode.Category.Value);
-        Assert.IsNull(result.MessageOrStructuredMessage.Message);
-        Assert.IsNotNull(result.MessageOrStructuredMessage.StructuredMessage);
-        Assert.AreEqual("001", result.MessageOrStructuredMessage.StructuredMessage.StructuredMessageType);
-        Assert.AreEqual("BVBA.BAKKER PIET                                                      ", result.MessageOrStructuredMessage.StructuredMessage.StructuredMessageFull);
-        Assert.IsEmpty(result.MessageOrStructuredMessage.StructuredMessage.Value);
+        Assert.That(result.SequenceNumber.Value, Is.EqualTo(1));
+        Assert.That(result.SequenceNumberDetail.Value, Is.EqualTo(1));
+        Assert.That(result.BankReference.Value, Is.EqualTo("0007500005482"));
+        Assert.That(result.TransactionCode.Type.Value, Is.EqualTo("0"));
+        Assert.That(result.TransactionCode.Family.Value, Is.EqualTo("04"));
+        Assert.That(result.TransactionCode.Operation.Value, Is.EqualTo("80"));
+        Assert.That(result.TransactionCode.Category.Value, Is.EqualTo("000"));
+        Assert.That(result.MessageOrStructuredMessage.Message, Is.Null);
+        Assert.That(result.MessageOrStructuredMessage.StructuredMessage, Is.Not.Null);
+        Assert.That(result.MessageOrStructuredMessage.StructuredMessage.StructuredMessageType, Is.EqualTo("001"));
+        Assert.That(result.MessageOrStructuredMessage.StructuredMessage.StructuredMessageFull, Is.EqualTo("BVBA.BAKKER PIET                                                      "));
+        Assert.That(result.MessageOrStructuredMessage.StructuredMessage.Value, Is.Empty);
     }
 
     [Test]
@@ -35,10 +37,10 @@ public class InformationPart1LineParserTest
         var parser = new InformationPart1LineParser();
         var sample = "31000100073403076534383000143  335370000ekeningING Plus BE12 3215 1548 2121 EUR Compte à vue BE25 3215 2158 2315             0 1";
 
-        Assert.IsTrue(parser.CanAcceptString(sample));
+        Assert.That(parser.CanAcceptString(sample), Is.True);
         var result = (InformationPart1Line)parser.Parse(sample);
 
-        Assert.AreEqual(1, result.SequenceNumber.Value);
-        Assert.AreEqual("ekeningING Plus BE12 3215 1548 2121 EUR Compte à vue BE25 3215 2158 2315 ", result.MessageOrStructuredMessage.Message.Value);
+        Assert.That(result.SequenceNumber.Value, Is.EqualTo(1));
+        Assert.That(result.MessageOrStructuredMessage.Message.Value, Is.EqualTo("ekeningING Plus BE12 3215 1548 2121 EUR Compte à vue BE25 3215 2158 2315 "));
     }
 }
